Fix warnings caused by los of integer precision

+++ mac/hid.c
@@ -818,7 +818,7 @@ static int set_report(hid_device *dev, IOHIDReportType type, const unsigned char
 					   data_to_send, length_to_send);
 
 		if (res == kIOReturnSuccess) {
-			return length;
+			return (int)length;
 		}
 		else
 			return -1;
@@ -848,7 +848,7 @@ static int return_data(hid_device *dev, unsigned char *data, size_t length)
 	dev->input_reports = rpt->next;
 	free(rpt->data);
 	free(rpt);
-	return len;
+	return (int)len;
 }
 
 static int cond_wait(const hid_device *dev, pthread_cond_t *cond, pthread_mutex_t *mutex)
@@ -1018,7 +1018,7 @@ int HID_API_EXPORT hid_get_feature_report(hid_device *dev, unsigned char *data, 
 	                           data[0], /* Report ID */
 	                           data, &len);
 	if (res == kIOReturnSuccess)
-		return len;
+		return (int)len;
 	else
 		return -1;
 }
